requirements.txt
discord.py==1.7.3
joblib==1.1.0
matplotlib==3.5.2
nltk==3.7
numpy==1.23.1
pandas==1.4.3
praw==7.6.0
requests==2.28.1
scikit_learn==1.1.1
textblob==0.17.1
